requirements.txt
cplex==12.9.0
networkx==2.6.3
numpy==1.21.6
scikit-learn==1.0.2
scipy==1.7.3
PyYAML==6.0
torch==1.13.0+cu116
torch-geometric==2.2.0
torch-scatter==2.1.1+pt113cu116
torch-sparse==0.6.17+pt113cu116